requirements.txt
numpy
pillow
docopt
tornado==4.5.3
requests
python-socketio
flask
eventlet
moviepy
pandas
imutils
opencv-python
pyserial
